Refactor : Spring Data Jpa를 사용한 refactor 진행중

+++ src/main/java/com/asac6c/reddit/dto/postDto/DraftDeleteRequestDto.java
@@ -13,6 +13,6 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(makeFinal = true, level = AccessLevel.PRIVATE)
 public class DraftDeleteRequestDto {
 
-    Integer postNo;
-    Integer userNo;
+    Long postNo;
+    Long userNo;
 }
