+++ src/main/java/com/asac6c/reddit/dto/postDto/PostCreateRequestDto.java
@@ -14,7 +14,7 @@ import lombok.experimental.NonFinal;
 @FieldDefaults(makeFinal = true, level = AccessLevel.PRIVATE)
 public class PostCreateRequestDto {
 
-    Integer userNo;
+    Long userNo;
     String postTitle;
     String postContent;
     boolean postDraft;
